ExceptionHandling final

+++ OSAHN6_HFT_202231.Logic/CoachLogic.cs
@@ -22,16 +22,21 @@ namespace OSAHN6_HFT_202231.Logic
 
         public void Create(Coach item)
         {
-            this.repo.Create(item);
+            if (item.CoachName.Length < 0 || item.CoachName.Length > 100) throw new FormatException();
+            if (repo.ReadAll().Select(x => x.CoachId).Contains(item.CoachId)) throw new FormatException();
+            if (item.Salary<=0) throw new FormatException();
+                this.repo.Create(item);
         }
 
         public void Delete(int id)
         {
+            if (!repo.ReadAll().Select(x => x.CoachId).Contains(id)) throw new FormatException();
             this.repo.Delete(id);
         }
 
         public Coach Read(int id)
         {
+            if (!repo.ReadAll().Select(x => x.CoachId).Contains(id)) throw new FormatException();
             return this.repo.Read(id);
         }
