+++ OSAHN6_HFT_202231.Logic/PlayerLogic.cs
@@ -20,16 +20,20 @@ namespace OSAHN6_HFT_202231.Logic
 
         public void Create(Player item)
         {
+            if (item.Name.Length <0 || item.Name.Length >100) throw new FormatException();
+            if (repo.ReadAll().Select(x => x.PlayerId).Contains(item.PlayerId)) throw new FormatException();
             this.repo.Create(item);
         }
 
         public void Delete(int id)
         {
+            if (!repo.ReadAll().Select(x => x.PlayerId).Contains(id)) throw new FormatException();
             this.repo.Delete(id);
         }
 
         public Player Read(int id)
         {
+            if (!repo.ReadAll().Select(x => x.PlayerId).Contains(id)) throw new FormatException();
             return this.repo.Read(id);
         }
 
@@ -40,6 +44,8 @@ namespace OSAHN6_HFT_202231.Logic
 
         public void Update(Player item)
         {
+            if (item.Name.Length < 0 || item.Name.Length > 100) throw new FormatException();
+            if (repo.ReadAll().Select(x => x.PlayerId).Contains(item.PlayerId)) throw new FormatException();
             this.repo.Update(item);
         }
     }
